AMP web push Chrome 62 notification API changes and other improvements (#11480)
* Disable widget when subscribing or unsubscribing

Until a subscription is complete or canceled, or until an unsubscription
is complete, disable the widget to prevent multiple taps, and then
enable the widget after the task is complete.
This commit changes the logging from development logging, which is
removed from the final build, to user logging, which is retained in the
final build. This should be helpful for debugging production sites.

* Allow localhost HTTP amp-web-push testing

For HTTP sites using web push through a push notification aggregator
service, testing the site in development required HTTPS for the initial
amp-web-push release. However, Chrome allows localhost and 127.0.0.1 to
be HTTP and still use powerful features reserved for secure origins.
Therefore, this commit also treats HTTP localhost or 127.0.0.1 as HTTPS
for easier testing.

* Smart AMP web push service worker URL match

Also fix tests for localhost / 127.0.0.1 HTTPS allowance.

* Chrome 62 permission changes

- Use a stored notification permission instead of the iframe's
  permission
- Don't open a subscription popup if the iframe's permission is already
  granted, just subscribe silently
- Handle notifications being blocked when opening the subscription popup

* Modify permission dialog structure, add tests, lint

* Whitelist use of localStorage, code review fixes

* Store notif perm earlier to allow popup

Store the notification permission before the user clicks the Subscribe
button, because the permission may currently be granted, and we can
avoid opening the popup in the future if the permission is granted.

The user might have previously subscribed and then cleared their
browser data without resetting their permission, and in this case we
don't have to.

We can't check the notification permission when the user clicks
Subscribe because waiting on promises after clicking buttons triggers
the browser's popup blocker. So we have to check and store the
permission before the user clicks Subscribe and use the stored result
when the user finally clicks Subscribe.

Also store the notification permission right after the user interacts
with the prompt, so we can store the result of blocking the prompt to be
used on next page load.