Add FormDataWrapper to read entries from FormData (#11495)
* Polyfill FormData#entries

This is a pre-requitesite for #11294.

Formdata#entries is not supported by IE, Edge and Safari, but is useful
for accessing content of the a `FormData` object. On IE, Edge and
Safari, the only available method on `FormData` is `append`, which means
that client code can't read from a `FormData`. Reading entries in a
`FormData` will be required by #11294 to serialize a `FormData` object
into a plain old JavaScript object so that it can be passed to
`postMessage()` (i.e., can be cloned using the structured clone
algorithm).

* Extract getFormAsObject_ to a shared place

This will be shared by a polyfill for FormData#entryArray, which reads
the content of the FormData and return a list of entries. To do that,
the polyfill actually needs to patch the FormData constructor and use
getFormAsObject_ to serialize the form element into a plain old object.

* Make FormData.entries polyfill return iterator

The polyfill can return an iterator by manually constructing an object
with next() function, as opposed to using a generator function which is
not supported by the current Babel configuration.

* Remove unnecessary arguments and blank lines

The arguments being passed are the default.

* Better handle installation isolation

Instead of storing the window where the polyfill is being installed onto
the function object, pass the window directly to the constructor
function by using `bind`. This ensures that separate installation for
distinct windows are isolated with different constructor function, each
having a reference to the native FormData constructor for the respective
window. Sharing state on the FormDataPolyfill function object across
window is problematic as you can't install the polyfills to multiple
windows.

Instead of copying the prototype to implicitly satisfy the precondition
check that prevents re-installation of the polyfill on the same window,
explicitly check that win.FormDataNative_ has been patched to prevent
re-entry of the `install` method on the same window.

For the tests, since `env.win.FormData` will be patched to the polyfill
function after calling `install`, `env.win.FormData` won't have anything
in the prototype. Instead, the tests should assert on the native
constructor's prototype.

Misc fix: Following AMP's convention of prefixing it with `opt_` and
giving it a default value (even if it's `undefined`).

* Correct iterator

1. Make the object literal use method shorthand
2. Return a `value: undefined` entry for the last iteration

* Rename identifiers and add notes for clarity

Some of the methods are just wrappers and not really polyfill
themselves. Some notes are added to clarify why subclassing wasn't used.

* Change FormData#entries polyfill to a wrapper

A wrapper is cleaner at the implementation side as it avoids the need
to mess with patching native constructors and having inconsistent
prototypes between polyfill constructor and native constructor.