Add disableScroll request to amp-viewer-integration
disableScroll/resetScroll are available on the Viewport but not
currently exposed to the viewer. scrollLock is already available, but
behaves different (preventing touchmove events) instead of setting
overflow:hidden, which will block *any* scrolling on the page, not just
within the iframe.

One of the new tests added is in a suite currently marked as 'skip', but
passes when run manually.