Use FormDataWrapper instead of native FormData
This is a pre-requitesite for #11294.

Formdata#entries is not supported by IE, Edge and Safari, but is useful
for accessing content of the a FormData object. On IE, Edge and
Safari, the only available method on FormData is append, which means
that client code can't read from a FormData. Reading entries in a
FormData will be required by #11294 to serialize a FormData object
into a plain old JavaScript object so that it can be passed to
postMessage() (i.e., can be cloned using the structured clone
algorithm).

Since polyfilling FormData#entries requires patching both the constructor
and the append method and can be ugly to implement, a new unified API
FormDataWrapper is added to provide a consistent way to deal with
FormData. Callers will be banned from calling new FormData(), and
will have to use new FormDataWrapper(). When a native FormData is
needed to be passed to the native fetch or XMLHttpRequest API, callers
would need to call formDataWrapper.getFormData().