Fix type checker errors
This fix the following errors thrown by the type checker when the
FormDataWrapper class starts to be used:

src/form-data-wrapper.js:79: ERROR - Property entries never defined on FormData
    if (!this.formData_.entries) {
                        ^^^^^^^

src/form-data-wrapper.js:101: ERROR - actual parameter 1 of Object.keys does not match formal parameter
found   : (Object|undefined)
required: Object
    Object.keys(this.fieldValues_).forEach(name => {
                ^^^^^^^^^^^^^^^^^

As well as fixing the nullability of the return value of `getFormData`.
It should really be non-nullable so that callers don't need to check
for null.

Also improved the type safety of the fieldValues_ variable to have the
key and value types.