Calculate correct documentHeight for short AMP docs (#12229)
* Use contentHeight for documentHeight messages

Handles the edge case of AMP documents which are shorter than their
containing viewport, which causes documentHeight to be scrollHeight,
which is the height of the viewport. Should be a no-op for docs which
are already taller than the viewport.

Re-implements changes made in
https://github.com/ampproject/amphtml/commit/ac741f6d448305570659d0eae3c11c9049ae0fef,
which were reverted before merging.

* Use HTML5 doctype for RealWinFixture

Non-HTML5 doctype has different behaviour for html/body element heights
in Safari.