Make our custom element implementation directly consumable by CE V1
The CustomElement V1 API is strictly coupled to ES6 classes inheriting from build-in `HTMLElement` or its subclasses. Alternatively, `Reflect.construct` can be used to make legacy constructors inherit from build-in classes.

Because we transpile ES6 classes, we never hit any of those 2 paths (Not an ES6 class and not using `Reflect.construct` in transpiled output). This is hidden by the polyfill by falling back to a V0 polyfill instead. As a consequence we depended on the polyfill even in browsers that implement custom elements.

This CL manually uses a legacy class with a `Reflect.construct` based constructor to make the resulting object directly consumable by `customElements.define`. Note, that while there is a way to make Babel produce similar code, there is no much way for closure compiler transpilation.

Additionally: The CE polyfill had a side effect and was included multiple times. Due to the polyfill's behavior this led to a forced CE V1 polyfill behavior (and it is also bad). We'll try to get this fixed upstream, but for now we always patch the polyfill (we had already done this for the prod build).