inabox: add window.AMP.inaboxUnregisterIframe(iframe)
The current API for inabox is that when you want to remove a frame you delete it
from ampInaboxIframes.  This means there isn't a way for inabox to run any
additional cleanup on frame removal, and currently there's a memory leak where
references to iframes (or their descendents) are maintained indefinitely in
InaboxMessagingHost.iframeMap_ and in the position observation listeners.

This changes the API from 'delete from a list of frames' to 'call a function',
which allows the function to do arbitrary cleanup like removing references.

I'm not that happy with the test, since it runs in a very mocked environment.
To test this manually I've added examples/inabox.host.memory.html which
repeatedly adds an iframe and removes it again.  This initially showed us as
leaking about 4MB of memory per refresh (desktop chrome; silly fake ad) but now
doesn't show any leaking.