inabox: fix memory leak
In #13795 we tried to fix a memory leak but missed that PositionObserver.observe(...) wasn't passing back the unregister function from Observable.add(), and so we weren't removing the listener.

After this change we still leak on refresh: with examples/inabox.host.memory.html I measure us as leaking 200k in 60 refreshes without this PR, vs 30k with it.  I'm still trying to figure out where else we leak, but getting the per-refresh leak down from 3k to 0.5k is still worth shipping.