🐛 Fix AmpShadowDoc#declareExtension_(string) behaviour when minified (#14896)
Previously marked as private, but used from the exported function `declareExtension` outside of the class (in the same module). This happens to work fine for `AmpDocSingle`, but property renaming screws up `AmpDocShadow` and the method breaks.

Discovered by running the `amp-next-page` example from minified AMP sources, which fails to install any extensions on the second page.