🐛 Extension race fix (#14998)
Fixes a race condition when all of the AMP JS is in browser cache except an extension that provides a service e.g. `amp-geo` and something calls the service before the extension registers.  This is particularly an issue for `amp-geo` because of it's short cache lifetime but it could hit any extension that provides a service.

It does this by detecting if there is a stubbed extension and if there is waiting for it (with a timeout).

Adds a timeout to `Extensions.waitForExtension()` defaults to 3 seconds which is the same as the timeout used for render blocking extensions, The timeout can be overridden with an `opt_timeout` param.

Adds `stubbedElementNames[]` to track what's been stubbed.

Updates tests that use `waitForExtension`

Adds unit test for `waitForExtension` timeout.

Fixes #14841