🐛Report if defaultView is null in preconnect (#15993)
This is an attempt at fixing #15883.

The current suspicion is that the setTimeout call inside
`CustomElement#preconnect` delays the call to
`BaseElement#preconnectCallback`. During this delay, the parent window
is destroyed (maybe the iframe is removed from the DOM?), causing
`element.ownerDocument.defaultView` to return `null`.

It's frankly very surprising that timers would fire after a window is
destroyed, so we're only guarding `#preconnectCallback` for now. We may
later extend this to all timers if it fixes the bug.