Install core services before installing builtins
Fixes #15759.

The core issue is a race condition between the startup chunking, and the
`buildCallback` of any of the builtin elements. In certain
circumstances, the `buildCallback` can actually be called sync after
installing the custom element. Then, it's a race between the adoption
chunk and `whenBodyAvailable`. If `whenBodyAvailable` wins, we get the
error.

So, let's just install the needed `extensions` service before installing
any of the builtins.