Update validator-amp-vk.protoascii
We can't use VK extension without VK components. But in our case, we have to use it without components. Because we don't know when VK posts will be in page, but we have to guess that it will be some day and we want to use this extension in our template without component. It's work in right way in YouTube, Facebook e.t.c. extensions.