Capture scroll events for ios-embed-sd
Scroll events don't bubble out of the shadow tree, sadly. Nor do they
bubble to the shadow root, ugh. So for this to be truly generic I would
have to install a MutaitonObserver on the shadow root, then listen for
scroll events on any child of the shadow root.

But that's too much work for now.