Fix error when URL macro whitelist is empty
Currently, if

<meta name="amp-variable-substitution-whitelist" content="">

is present in the doc and the doc has, for example, an amp-list, then
the runtime will throw this error:

Cannot set property 'name' of undefined

This occurs here:

https://github.com/ampproject/amphtml/blob/8e799b53c96e84840e73b7d1b5c9935536c3d1a5/src/service/url-expander/expander.js#L133

This is because currently VariableSource.prototype.getExpr returns the
following regex when the URL macro whitelist is empty:

\$?()

Note the empty parentheses. This regex actually matches the empty
string, so when tested against any string, the match result will be an
array of empty strings of size n+1, where n is the size of the tested
string.

To fix that, we need to return a regex that matches nothing when the
whitelist is empty.