✨ Support (open) shadow roots in layers (#17774)
* Support (open) shadow roots in layers

* Update comments

* Capture scroll events for ios-embed-sd

Scroll events don't bubble out of the shadow tree, sadly. Nor do they
bubble to the shadow root, ugh. So for this to be truly generic I would
have to install a MutaitonObserver on the shadow root, then listen for
scroll events on any child of the shadow root.

But that's too much work for now.

* Cleanup listening

* fix name

* Comments

* Use assignedSlot

* Remove unused import

* Private annotations

* Add tests

* Bump

* Bump 2

https://travis-ci.org/ampproject/amphtml/jobs/422756893

* skipTest