Make sure template rendering works when detached.
Change `viewerCanRenderTemplates` to use a cached reference to the
viewer. Since the template element can later be detached, we need to
make sure we already have the service reference. We cannot look up the
service when the template is detached.

We cannot have `<amp-access>` keep the template in the DOM since that
will impact CSS selectors (e.g. :first-child or .foo + .bar) as we
will be altering the DOM structure from the current state.