Layers: prevent multiple references to children layouts
Parent layers are always aware of there children, but children may not
be aware of there parent layers. When a child finds its parent, it adds
itself to the parent (ie, when a new element is added to the DOM, it
adds itself to the parent).

In some circumstances, this lead to multiple references of child layouts.
When FIE destroyed itself, the first reference was being destroyed, but
that's it. Those extra references were kept alive, and the next time
they were measured an error is thrown because the child layouts window
context is destroyed (preventing accurate measurements).