Renames filterSplice to remove
I noticed no one is using the return value from `filterSplice` (the
spliced out elements). So I removed that.

But then I noticed this isn't really like splice anymore, so I renamed
it to `filter`. But that's too close to `Array.p.filter` (and this is
destructive), so I renamed to `remove` to highlight the difference.