Unset inline display:none for nodisplay elements (#18419)
Server Side Rendering currently injects an inline `display: none` into
`nodisplay` AMP elements. Now that we've [switched](https://github.com/ampproject/amphtml/pull/17988)
to using the `hidden` attribute, we'll need to unset this inline style
when applying the static layouts to the element.

This is temporary until we can update SSR to inject a `hidden` attribute
instead of the inline style.