Layers: fix Resource detection of measurement changes (#18553)
Layers tries to eagerly remeasure whatever is necessary. In some cases, that means that we could remeasure a child element _before_ necessary, even though Resources will still think it needs remeasure.

In most cases, this would be fine. But, because we were using the latest value of `getPageLayoutBox`, it would fail to detect that measurements changed. Instead, we need to keep track of the last time we meausred, and compare against that.