AMP-Cache-Transform: specify version syntax.
This specifies a `v` parameter as a list of integer ranges, and
specifies how servers and caching proxies should handle it in both
requests and responses.

It does not specify the values of those integers; it is expected that
the canonical source of those values will be the transformer library in
github.com/ampproject/amppackager.

Fixes #18046.