Use chunk function to scan the document.
Chunk will delay the execution of the document scaning to when the browser is idle. The purpose of this change is to avoid any potential performance issue that could affect the user experience when loading the page.

At this point it hasn't been proven that amp-skimlinks could create such  slowdown but it feels safer to wrap it inside the chunk function for now. Since this is not the ideal solution (we could potentially lose clicks and page impressions), it would be great make more measurement and optimization and remove it once we feel confident that amp-skimlinks can run without risking to interfer with the user experience.