Allow forcing the Custom Elements V1 Polyfill
Omitting the `ctor` when installing the CE V1 polyfill will now force
the full polyfill installation, since we can't properly test if Native
Classes are being used (necessary for proper native support).

This will allow us to launch the new polyfill without native CE support,
which is the riskiest part of the transition.