convert some `dev().assert` calls to `user().assert` calls
When we remove the method calls (assert method of any type) and add a
straight type cast annotation, closure seems to infer that the values
for these instances are not of that exact type (meaning they could be
null). To prove this, I am switching the calls over to `user` type calls
so that we don't remove the method call and add an error message that I
can trace better.