Stop amp-sidebar from forwarding hotkeys to viewer (#20129)
The keyboard event handler in the AMP viewer extension stops forwarding
keyboard events to the viewer if the event is marked as "default
prevented". For hotkeys used by amp-sidebar (i.e., the ESC key), since
they are already consumed by the AMP component they should not be
forwarded to the viewer. This prevents double-triggering a keyboard
shortcut on the viewer side.

For more background, see #18437.

`b/117567983`