✅Amp-skimlinks end to end tests (#20211)
Add end to end tests fo amp-skimlinks extension.

#### Notes:
In order to test amp-skimlinks script, we introduced a new **optional** JSON config to overwrite the request endpoints used by amp-skimlinks. The config (used only for the end to end tests) can be inserted inside the `<amp-skimlinks>` tag as follow:

```html
<amp-skimlinks
    layout="nodisplay"
    publisher-code="123X123"
    tracking="true"
>
    <script type="application/json">
    {
        "pageTrackingUrl": "${pageTrackingUrl}",
        "linksTrackingUrl": "${linksTrackingUrl}",
        "nonAffiliateTrackingUrl": "${nonAffiliateTrackingUrl}",
        "waypointUrl": "${waypointUrl}"
    }
    </script>
</amp-skimlinks>
```

This allows us to overwrite the API endpoints to use the "request-bank" instead, so we can later validate if an API call has happened as expected or not.