Add RC binary type (#20463)
We're releasing an RC build, which will be 100% the same as the code
that _wil be released_ to production next week. The canary type sticks
around, but it is (as it has always been) an experimental configuration
of next week's code.

Proper experiment comparisons should be done `canary <-> rc`, and `rc
<-> control`. Comparing `canary <-> control` will not give the correct
experiment configurations, because it tests experimental code _and_
experimental configurations.