Fix amp-sidebar keyboard event forwarding
It turns out that the close_() function may be a no-op if the sidebar
has already been closed, in which case we should not prevent the
default.

Currently, just by having an amp-sidebar in the AMP document, and
clicking on a blank region in the document would prevent the default and
stop forwarding the key event to the viewer. This is because the
keyboard event listener is attached onto the `document` object. We still
need to forward the keyboard event when the sidebar is already closed.