Introduce a secure way to embed sandboxed AMP
For an AMP document embedded inside a sandbox iframe, the origin of the
document would be "null", which defeats the purpose of an origin check.
An attacker could simply create a sandboxed, malicious iframe (therefore
having null origin), walk on the DOM frame tree to find a reference to
the viewer iframe (this is not constrained bu the same origin policy),
and then send postMessage() calls to the viewer frame and pass the
viewer's origin checks, if any.

The viewer could also check the source of the message to be a legitimate
AMP iframe window, but the attacker could bypass that by navigating the
legitimate AMP iframe window away to a malicious document. Recent
browsers have banned this kind of attack, but it's tricky to rely on it.

To prevent the above attack in a null origin AMP document, the viewer
should include this token in an init parameter, either in the `src` or
`name` attribute of the iframe, and then verify that this token is
included in all the messages sent from AMP to the viewer. The attacker
would not be able to steal this token under the same origin policy,
because the token is inside the viewer document at a different origin
and the attacker can't access it.