Fix check-type
AmpViewerMessage is defined as a @struct, which can't be modified after
its construction. I also can't use the square bracket notation to mutate
it. I also don't want to modify the struct definition to allow the
`messagingToken` field, since I want to conditionally set it on the
message depending on whether it's undefined (as opposed to having
something like `messagingToken: undefined` in the object).