Remove side-slide-count functionality from amp-base-carousel.
This feature has some issues, and is no longer needed since Chrome is
implementing `scroll-snap-stop`, which allows us to tell the browser to
stop on every snap point rather than trying to force it by hiding
slides.

This was never documented like the other attributes, so no docs changes
are needed.