Layers: guard against repeated cross-origin accesses
Part of the layers algorithm is to traverse out of FIE frames to the
parent AMP doc. To do this, when we hit the document node, we test to
see if we can access the document's window's `frameElement`. However,
if the parent window is cross-origin, this will cause an error.

We actually guard against this error, so there is real issue. But these
access are still logged in dev tools, which gets really annoying. This
is a nice-to-have change so that we only attempt getting the
`frameElement` once, and cache the result.