✨ mraid: change default for fallback, and simplify configuration (#21738)
* mraid: change default for fallback, and simplify configuration

Initially the design for amp-mraid was that there were multiple situations in which it would need to decide whether to fall back to standard web apis, and we wanted to allow the author to choose how amp-mraid should handle that situation.  It's now clear that this was too complicated, and the only decision amp-mraid needs the author to make is how it should handle the case where it doesn't look like its running in a mobile app.

Additionally, when a creative author decides to include mraid support in a creative they're generally going to want the fallback behavior.  They don't know what environment they will be serving into, and want reasonable behavior everywhere.  Make fallback the default.  Ad networks that know at serving time that they're serving to mobile apps can use no-fallback to ensure that an error with mraid.js doesn't result in incorrect usage of web apis.

* remove stray console.log