Lightbox captions changes.
- Refactor lightbox captions into a separate file.
- Fix downwards movement at the start of the captions expanding by
having consistent padding both before and after the expansion.
- Allow opening lightbox with the captions already expanded via an
action.
- Apply captions fading mask when switching slides, which is less
jarring when the new slide's captions are overflowing.
- Keep the close and gallery view buttons visible when the caption is
expanded. This is important when opening when captions already expanded.
  * Added a fade out effect to the top of the text, so it does not clash
  with the buttons.
- Change the caption fade out mask to use `mask-image` instead of a
gradient. This works well when the captions are overlaying the image
(e.g. if you enter pan-zoom mode or the image is in portrait).
  * Add a text-shadow around the captions as a replacement, so when the
  captions are on top of an image, they are still readable. This also
  improves the contrast when the captions are expanded.