Various changes for amp-base-carousel.
- Only go to the next item when the carousel arrow itself is clicked
rather than the element that contains the arrow (which is full height).
- Disable looping when there are less than 3x slides compared to the
visible count, since the UI does not behave  correctly. The threshold for this
can be reduced, but requires more work.
- Add an `allowWrap` option to `advance`. This is used to prevent the
next/prev arrows from looping (when looping is disabled) if they are quickly
clicked before they are disabled. Autoadvance is still allowed to wrap
when looping is disabled.
- Use the `scrollend` event, when available, to immediately restore the
scrolling reference point. This ne feature allows us to determine when
scrolling ends directly instead of waiting for a gap after a scroll
event.
  * We currently wait 200ms after the last scroll, which appears to be
  the minimum value that works for Android / iOS.
  * This greatly reduces the chance that a user could hit the end point
  of a looping carousel when swiping quickly and continuously.
  * https://www.chromestatus.com/feature/5650553247891456
- Fix the buttons not being disabled properly when looping visible count is
>= 2 or using mixed-lengths.
  * Check that the scroll position is at the end, since we may never
  have the last slide as the current index.
- Fix race condition when switching between looping and non-looping
during a scree resize, where the current index would not be updated correctly.
- Fix scroll position calculation (for restoring later), which was
incorrectly adding `containerStart` instead of subtracting. This causes
the position to be resotred incorrectly when the carousel is not aligned
at the edge of the window.
- Fix the movement of slides to the left/right when there is a visible
count other than one. Previously, too many slides would be moved to the
right, and not enough moved to the left.