Fix quadratic remeasurements during initial page load
By forcing remeasure during resource measure, we were invalidating both the element and it's entire parent layer. This had the effect of remeasuring every element on the page when the parent layer is the root scroller.

`Resources` calls this measurement during it's initial page rendering (in a loop on every `Resource`). That gets us n<sup>2</sup> calls to `getBoundingClientRect()` and `clientWidth`/`clientHeight`, which slowed everything down.

Thankfully, there was only one caller that actually needed a forced remeasurement.