Dispatch VALUE_CHANGE on `amp-bind` value changes
This introduces a new `AmpEvents.VALUE_CHANGE` event. It is dispatched
when `amp-bind` mutates form input values, specifically:
`<input [value]>`, `<textarea [text]>`, `<input [checked]>`, and
`<option [selected]>`.

A custom event is chosen over a mutation observer, because binding
`[text]` on `textarea` does not change the DOM (and it's expected), thus
a mutation observer won't pick up the change. The same applies to
`input` elements, when we eventually have two separate bindings for
`[value]` and `[defaultValue]`.