🚀 Layers: Fix quadratic remeasurements during initial page load (#22809)
* Fix quadratic remeasurements during initial page load

By forcing remeasure during resource measure, we were invalidating both the element and it's entire parent layer. This had the effect of remeasuring every element on the page when the parent layer is the root scroller.

`Resources` calls this measurement during it's initial page rendering (in a loop on every `Resource`). That gets us n<sup>2</sup> calls to `getBoundingClientRect()` and `clientWidth`/`clientHeight`, which slowed everything down.

Thankfully, there was only one caller that actually needed a forced remeasurement.

* Bust layers measurement when requetsing a remeasure

* Optimize remeasuring children layouts

If the containing layer didn't change, there's no need to remeasure the children.

* Only transfer children if we know the parent layer

* When a new child layer is added, transfer ownership

* lint

* Remove remeasure optimization

* Fix inabox tests

* Skip flaky test