Dispatch VALUE_CHANGE on `amp-bind` value changes (#22840)
* Dispatch VALUE_CHANGE on `amp-bind` value changes

This introduces a new `AmpEvents.VALUE_CHANGE` event. It is dispatched
when `amp-bind` mutates form input values, specifically:
`<input [value]>`, `<textarea [text]>`, `<input [checked]>`, and
`<option [selected]>`.

A custom event is chosen over a mutation observer, because binding
`[text]` on `textarea` does not change the DOM (and it's expected), thus
a mutation observer won't pick up the change. The same applies to
`input` elements, when we eventually have two separate bindings for
`[value]` and `[defaultValue]`.

* more elaborate test names

* rename event to FORM_VALUE_CHANGE

* dispatch event at <select> when <option> changes

* extract FORM_VALUE_CHANGE_EVENT_ARGUMENTS in test

* dont use `closest` & move everything into a helper

* refactor to use a lookup map instead

* avoid a linear `includes` search