Restrict focus / screenreaders to sidebar when it is open.
This change only applies to `<amp-sidebar>`. Changes for
`<amp-lightbox>`, `<amp-image-lightbox>` and `<amp-lightbox-gallery>`
will be made as a followup.

- Add aria-hidden to all content outside of the sidebar when the sidebar
is open. Trying to activate things outside the sidebar will not work
correctly due to the overlay.
- Make all potentially focusable elements have a tabindex of -1,
preventing them from gaining focus. This is different from other
implementations that do focus trapping, in that it allows focus to
return to the browser.
- Make sure to focus something within the sidebar for VoiceOver, which
does not move the screen reader's focus over when focusing the whole
sidebar.
- Add a close button to the start of the sidebar if the page author did
not specify a close button. This is in addition to the screen-reader
close button we add to the end of the sidebar.