🚀 Layers: Lazily remeasure child layouts (#22992)
* Layers: Lazily remeasure child layouts

This is attempt 2 to fix the quadratic remeasurements described in #22809.

This changes the layers code so that child layouts are remeasured as needed, not eagerly when the parent layer is remeasured. Still, causing any mutation inside a layer causes all layouts inside that layer to be invalidated. But since we no longer need to remeasure everything in that layer to get the measurements of a single layout, this leads to significantly less remeasures overall.

* Fix double-references test