TEST: createElement instead of insertAdjacentHTML (#23454)
The `defaultValue` property returns "the default value as originally
specified in the HTML that created this object."
https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement#Properties

Test elements for `isFieldDefault` were originally created with
`insertAdjacentHTML` to ensure their `defaultValue` properties is
generated correctly. It turns out `setAttribute` actually modifies the
element's HTML definition, and thus could update the `defaultValue`
property.

This change migrates the test to use the preferred `createElement` and
`setAttribute` to set up test elements.