Add jsonConfiguration transform
The `jsonConfiguration` transform is a static helper that takes large configuration objects and converts them to a JSON string representation at build time. This improves load time and memory of the config object.

We keep the configuration as a JS Object in the source code to make development easier. We can add comments as necessary (not possible in JSON), can use traililng comments, get nice syntax highlighting, etc.

Only true JSON  configurations can be represented this way. Any dynamic behavior inside the configuration is forbidden.

The actual `jsonConfiguration` funciton will never be included in the builds, it's purpose is to signal to the babel transform. It exists only in development and tests.

```js
// Input
const config = jsonConfiguration({foo: 'bar'});

// Output
const config = JSON.parse('{"foo": "bar"}');
```