Overhaul Chunks implementation and avoid macro tasks before body is visible.
This change got much bigger than planned. On top of a functional change there is one major refactoring: We now never schedule another execution until after the current one is done. This ensures that we don't over-schedule which effectively voids the point of multiple scheduling strategies. In a way this is a risky change, because previously we always ensured that we'd schedule an execution if we added a task. Now this relies on the state machine being correct.

Unskips the tests for long tasks.

Functional change: Avoids scheduling macro-tasks if the body is still hidden, because we want to get out of this state ASAP. Improves FCP with the experiment on by about 20%.

Related to #23618