Create amp-carousel 0.2 based on amp-base-carousel (#23393)
The main goal of this PR is to create a mostly API compatible update to `<amp-carousel>`, addressing numerous bugs. This is not strictly API compatible due to how the arrow buttons are positioned to allow automatic flipping in RTL.

This uses the same DOM structure as amp-carousel as far as slides and their classes are concerned for both `type="slides"` and `type="carousel"`. This maintains the API of the existing carousel. This does not add support for media queries in attributes, handling updating the slides or any other features of `<amp-base-carousel>`.

- Fix a bad usage of `clamp`
- Rename some base carousel classes to not clash
- Make scrollbar hiding optional, as `type="carousel"` does not hide the scrollbar
- Use `setOwner` for managing when to schedule layout for children
- Use `IntersectionObserver` to determine when to layout/unlayout/pause/resume slides
  * This is mostly important for `type="carousel"`, where slides can be of mixed length, and we cannot rely on the current index to tell when to actually layout adjacent slides
  * If IntersectionObserver is not supported, this currently skips `setOwner` and just relies on the runtime
- Fix arrow button hiding when at the start/end for RTL
- Fix bug where if you touch/wheel scrolled exactly when the scroll position was being reset, the current index would not be reported correctly